+++ Model Classes/Patch.m
@@ -1,5 +1,5 @@
 #import "Patch.h"
-#import <PYMIDI/PYMIDI.h>
+#import "PYMIDI/PYMIDI.h"
 
 
 @interface Patch (private)
@@ -493,7 +493,7 @@ BOOL isStatusByte (Byte b)		{ return b >= 0x80 && b < 0xF8; }
 BOOL isRealtimeByte (Byte b)	{ return b >= 0xF8; }
 
 
-unsigned int
+long
 midiPacketListSize (const MIDIPacketList* packetList)
 {
     const MIDIPacket*	packet;
